Relocate every source file into a more suitable folder
