+++ package.json
@@ -4,9 +4,9 @@
   "scripts": {
     "tsconfig": "echo '{ \"extends\": \"@tsconfig/recommended/tsconfig.json\" }' > tsconfig.json",
     "build": "npm run tsconfig && npm run build-pre && npm run build-main && npm run build-post",
-    "build-pre": "ncc build --target es2019 src/pre.ts && mv dist/index.js dist/pre.js",
-    "build-main": "ncc build --target es2019 src/main.ts && mv dist/index.js dist/main.js",
-    "build-post": "ncc build --target es2019 src/post.ts && mv dist/index.js dist/post.js",
+    "build-pre": "ncc build --target es2019 src/action/pre.ts && mv dist/index.js dist/pre.js",
+    "build-main": "ncc build --target es2019 src/action/main.ts && mv dist/index.js dist/main.js",
+    "build-post": "ncc build --target es2019 src/action/post.ts && mv dist/index.js dist/post.js",
     "contributors:add": "all-contributors add",
     "contributors:generate": "all-contributors generate",
     "contributors:check": "all-contributors check",
@@ -62,7 +62,7 @@
   },
   "ava": {
     "files": [
-      "src/*.test.ts"
+      "src/**/*.test.ts"
     ],
     "extensions": [
       "ts"
