+++ src/action/main.ts
@@ -2,13 +2,13 @@ import fs from 'fs'
 import process from 'process'
 import * as core from '@actions/core'
 import {getOctokit} from '@actions/github'
-import * as coursier from './coursier'
-import {type Files} from './files'
-import {GitHub} from './github'
-import {Input} from './input'
-import {type Logger} from './logger'
-import {nonEmpty, NonEmptyString} from './types'
-import * as workspace from './workspace'
+import * as coursier from '../modules/coursier'
+import {type Files} from '../core/files'
+import {GitHub} from '../modules/github'
+import {Input} from '../modules/input'
+import {type Logger} from '../core/logger'
+import {nonEmpty, NonEmptyString} from '../core/types'
+import * as workspace from '../modules/workspace'
 
 /**
  * Runs the action main code. In order it will do the following:
