+++ src/modules/coursier.ts
@@ -5,7 +5,7 @@ import * as core from '@actions/core'
 import * as tc from '@actions/tool-cache'
 import * as io from '@actions/io'
 import * as exec from '@actions/exec'
-import {type NonEmptyString} from './types'
+import {type NonEmptyString} from '../core/types'
 
 /**
  * Install `coursier` and add its executable to the `PATH`.
